Trigger geopackage generation fixes and refactoring * moving from cron.py to tasks.py * celery beat configuration on docker-compose.yml * geopackage tasks configured to use default queue
