--- src/planscape/planning/tests/test_cron.py
@@ -1,61 +1,0 @@
-from unittest import mock
-
-from django.test import TestCase
-from planning.models import GeoPackageStatus, ScenarioResultStatus
-from planning.tests.factories import ScenarioFactory
-from planning.cron import trigger_geopackage_generation
-
-
-class TriggerGeopackageGenerationCronJobTestCase(TestCase):
-    @mock.patch("planning.cron.async_generate_scenario_geopackage.delay")
-    def test_trigger_geopackage_generation_cron_job(self, mock_async_generate):
-        scenario = ScenarioFactory.create(
-            result_status=ScenarioResultStatus.SUCCESS,
-            geopackage_status=GeoPackageStatus.PENDING,
-        )
-
-        trigger_geopackage_generation()
-
-        mock_async_generate.assert_called_once_with(scenario.pk)
-
-    @mock.patch("planning.cron.async_generate_scenario_geopackage.delay")
-    def test_trigger_geopackage_generation_no_pending(self, mock_async_generate):
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.SUCCESS,
-            geopackage_status=GeoPackageStatus.PROCESSING,
-        )
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.SUCCESS,
-            geopackage_status=GeoPackageStatus.SUCCEEDED,
-        )
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.SUCCESS,
-            geopackage_status=GeoPackageStatus.FAILED,
-        )
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.SUCCESS,
-            geopackage_status=None,
-        )
-
-        trigger_geopackage_generation()
-        mock_async_generate.assert_not_called()
-
-    @mock.patch("planning.cron.async_generate_scenario_geopackage.delay")
-    def test_trigger_geopackage_generation_result_status_not_success(
-        self, mock_async_generate
-    ):
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.PENDING,
-            geopackage_status=GeoPackageStatus.PENDING,
-        )
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.RUNNING,
-            geopackage_status=GeoPackageStatus.PENDING,
-        )
-        ScenarioFactory.create(
-            result_status=ScenarioResultStatus.FAILURE,
-            geopackage_status=GeoPackageStatus.PENDING,
-        )
-
-        trigger_geopackage_generation()
-        mock_async_generate.assert_not_called()
